# src/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "LinkIO API",
      "version": "1.0.0",
      "description": "API for digital dentistry",
      "contact": {
        "name": "Caio Mikoszewski",
        "email": "caio.mikoszewski@arrowdigital.group"
      },
      "license": {
        "name": "AllRights-Attribution-NonCommercial-NoDerivatives 4.1",
        "url": "https://github.com/arrow-digital/license"
      }
    },
    "servers": [
      {
        "url": "https://vastly-cheerful-cougar.ngrok-free.app",
        "description": "NGROK tunnel server"
      },
      {
        "url": "http://localhost:8080",
        "description": "Local server"
      }
    ],
    "paths": {
...
      "/api/models/lab": {
        "get": {
          "summary": "Get all labs",
          "tags": [
            "Lab"
          ],
          "responses": {
            "200": {
              "description": "List of labs",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Lab"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "post": {
          "summary": "Create a new lab",
          "tags": [
            "Lab"
          ],
          "requestBody": {
            "description": "Lab object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lab"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Created lab",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Lab"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/lab/{id}": {
        "get": {
          "summary": "Get a lab by ID",
          "tags": [
            "Lab"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the lab to retrieve",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Lab object",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Lab"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "put": {
          "summary": "Update a lab by ID",
          "tags": [
            "Lab"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the lab to update",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "description": "Lab object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lab"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Updated lab",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "delete": {
          "summary": "Delete a lab by ID",
          "tags": [
            "Lab"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the lab to delete",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Deleted lab",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/inventory": {
        "get": {
          "summary": "Get all inventories",
          "tags": [
            "Inventory"
          ],
          "responses": {
            "200": {
              "description": "List of inventories",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Inventory"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "post": {
          "summary": "Create a new inventory",
          "tags": [
            "Inventory"
          ],
          "requestBody": {
            "description": "Inventory object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventory"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Created inventory",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Inventory"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/inventory/{id}": {
        "get": {
          "summary": "Get a inventory by ID",
          "tags": [
            "Inventory"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the inventory to retrieve",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Inventory object",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Inventory"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "put": {
          "summary": "Update a inventory by ID",
          "tags": [
            "Inventory"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the inventory to update",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "description": "Inventory object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventory"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Updated inventory",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "delete": {
          "summary": "Delete a inventory by ID",
          "tags": [
            "Inventory"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the inventory to delete",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Deleted inventory",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/order": {
        "get": {
          "summary": "Get all orders",
          "tags": [
            "Order"
          ],
          "responses": {
            "200": {
              "description": "List of orders",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "post": {
          "summary": "Create a new order",
          "tags": [
            "Order"
          ],
          "requestBody": {
            "description": "Order object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Created order",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/order/{id}": {
        "get": {
          "summary": "Get a order by ID",
          "tags": [
            "Order"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the order to retrieve",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Order object",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "put": {
          "summary": "Update a order by ID",
          "tags": [
            "Order"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the order to update",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "description": "Order object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Updated order",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "delete": {
          "summary": "Delete a order by ID",
          "tags": [
            "Order"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the order to delete",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Deleted order",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Inventory": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "amount": {
                    "description": "must be a number and is optional",
                    "type": "number"
                },
                "clinic": {
                    "description": "must be an objectId from the clinic collection and is optional",
                    "type": "string"
                },
                "description": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "lab": {
                    "description": "must be an objectId from the lab collection and is optional",
                    "type": "string"
                },
                "name": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "type": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "value": {
                    "description": "must be a number and is optional",
                    "type": "number"
                }
            },
            "title": "Inventory"
        },
        "Lab": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "address": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "clinics": {
                    "description": "must be an array of objectIds from the clinics collection and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "collaborators": {
                    "description": "must be an array of objectIds from the users collection and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "email": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "inventory": {
                    "description": "must be an array of objectIds from the inventory collection and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "name": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "orders": {
                    "description": "must be an array of objectIds from the orders collection and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "phone": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "revenue": {
                    "description": "must be an array of objects and is required",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "services": {
                    "description": "must be an array of objectIds from the services collection and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "taxid": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "users": {
                    "description": "must be an array of objectIds from the users collection and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                }
            },
            "required": [
                "name",
                "revenue"
            ],
            "title": "Lab"
        },
        "Order": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "clinic": {
                    "description": "must be an objectId from the clinic collection and is optional",
                    "type": "string"
                },
                "comments": {
                    "description": "must be an array of objects and is required",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "createdAt": {
                    "description": "must be a date and is required",
                    "format": "date-time",
                    "type": "string"
                },
                "dentist": {
                    "description": "must be an objectId from the dentists array in the clinic collection and is optional",
                    "type": "string"
                },
                "description": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "expiresAt": {
                    "description": "must be a date and is required",
                    "format": "date-time",
                    "type": "string"
                },
                "lab": {
                    "description": "must be an objectId from the lab collection and is optional",
                    "type": "string"
                },
                "patient": {
                    "description": "must be an objectId from the patients array in the clinic collection and is optional",
                    "type": "string"
                },
                "services": {
                    "description": "must be an array of objects and is required",
                    "items": {
                        "properties": {
                            "_id": {
                                "description": "must be an objectId and is optional",
                                "type": "string"
                            },
                            "collaborator": {
                                "description": "must be an objectId from the collaborator collection and is optional",
                                "type": "string"
                            },
                            "createdAt": {
                                "description": "must be a date and is required",
                                "format": "date-time",
                                "type": "string"
                            },
                            "createdBy": {
                                "description": "must be an objectId from the user collection and is optional",
                                "type": "string"
                            },
                            "discount": {
                                "description": "must be a number and is required",
                                "type": "number"
                            },
                            "expiresAt": {
                                "description": "must be a date and is required",
                                "format": "date-time",
                                "type": "string"
                            },
                            "finalvalue": {
                                "description": "must be a number and is required",
                                "type": "number"
                            },
                            "jaw": {
                                "description": "must be an array of numbers and is optional",
                                "items": {
                                    "type": "number"
                                },
                                "type": "array"
                            },
                            "service": {
                                "description": "must be an objectId from the service collection and is optional",
                                "type": "string"
                            },
                            "state": {
                                "description": "must be a string and is required",
                                "type": "string"
                            },
                            "teeth": {
                                "description": "must be an array of numbers and is optional",
                                "items": {
                                    "type": "number"
                                },
                                "type": "array"
                            }
                        },
                        "required": [
                            "createdAt",
                            "expiresAt",
                            "finalvalue",
                            "state",
                            "discount"
                        ],
                        "type": "object"
                    },
                    "type": "array"
                },
                "state": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "status": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "tag": {
                    "description": "must be an array of strings and is required",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                }
            },
            "required": [
                "status",
                "services",
                "description",
                "state",
                "createdAt",
                "expiresAt",
                "comments",
                "tag"
            ],
            "title": "Order"
        },
...
      }
    }
}